+++ Office365/CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2024-11-05 - 2.18.3
+
+### Fixed
+
+- Change the shutdown the connector
+
 ## 2024-11-04 - 2.18.2
 
 ### Fixed
 
-- Chnage the way to use the asyncio event loop
+- Change the way to use the asyncio event loop
 
 ## 2024-11-04 - 2.18.1